+++ src/torchoutil/utils/collections.py
@@ -7,55 +7,98 @@ from typing import (
     Dict,
     Iterable,
     List,
+    Literal,
     Mapping,
     Optional,
     Sequence,
     Tuple,
     TypeVar,
     Union,
+    overload,
 )
 
 from torchoutil.utils.type_checks import is_mapping_str_any
 
+K = TypeVar("K")
 T = TypeVar("T")
 U = TypeVar("U")
+V = TypeVar("V")
+W = TypeVar("W")
+
+KEY_MODES = ("same", "intersect", "union")
+KeyMode = Literal["intersect", "same", "union"]
+
+
+@overload
+def list_dict_to_dict_list(
+    lst: Sequence[Mapping[K, V]],
+    key_mode: Literal["intersect", "same"],
+    default_val: Any = None,
+) -> Dict[K, List[V]]:
+    ...
+
+
+@overload
+def list_dict_to_dict_list(
+    lst: Sequence[Mapping[K, V]],
+    key_mode: Literal["union"] = "union",
+    default_val: W = None,
+) -> Dict[K, List[Union[V, W]]]:
+    ...
 
 
 def list_dict_to_dict_list(
-    lst: Sequence[Mapping[str, T]],
-    default_val: U = None,
-    error_on_missing_key: bool = False,
-) -> Dict[str, List[Union[T, U]]]:
-    """Convert a list of dicts to a dict of lists.
+    lst: Sequence[Mapping[K, V]],
+    key_mode: KeyMode = "union",
+    default_val: W = None,
+) -> Dict[K, List[Union[V, W]]]:
+    """Convert list of dicts to dict of lists.
 
-    Example 1
-    ----------
-    >>> lst = [{'a': 1, 'b': 2}, {'a': 4, 'b': 3, 'c': 5}]
-    >>> output = list_dict_to_dict_list(lst, default_val=0)
-    {'a': [1, 4], 'b': [2, 3], 'c': [0, 5]}
+    Args:
+        lst: The list of dict to merge.
+        key_mode: Can be "same" or "intersect".
+            If "same", all the dictionaries must contains the same keys otherwise a ValueError will be raised.
+            If "intersect", only the intersection of all keys will be used in output.
+            If "union", the output dict will contains the union of all keys, and the missing value will use the argument default_val.
+        default_val: Default value of an element when key_mode is "union". defaults to None.
     """
-    if len(lst) == 0:
+    if len(lst) <= 0:
         return {}
 
-    if error_on_missing_key:
-        keys = set(lst[0])
-        for dic in lst:
-            if keys != set(dic.keys()):
-                raise ValueError(
-                    f"Invalid dict keys for list_dict_to_dict_list. (found {keys} and {dic.keys()})"
-                )
+    keys = set(lst[0].keys())
+    if key_mode == "same":
+        if not all(keys == set(item.keys()) for item in lst[1:]):
+            raise ValueError("Invalid keys for batch.")
+    elif key_mode == "intersect":
+        keys = intersect_lists([item.keys() for item in lst])
+    elif key_mode == "union":
+        keys = union_lists([item.keys() for item in lst])
+    else:
+        raise ValueError(
+            f"Invalid argument key_mode={key_mode}. (expected one of {KEY_MODES})"
+        )
 
-    keys = {}
-    for dic in lst:
-        keys.update(dict.fromkeys(dic.keys()))
+    return {key: [item.get(key, default_val) for item in lst] for key in keys}
 
-    out = {
-        key: [
-            lst[i][key] if key in lst[i].keys() else default_val
-            for i in range(len(lst))
-        ]
-        for key in keys
-    }
+
+def intersect_lists(lst_of_lst: Sequence[Iterable[T]]) -> List[T]:
+    """Performs intersection of elements in lists (like set intersection), but keep their original order."""
+    if len(lst_of_lst) <= 0:
+        return []
+    out = list(dict.fromkeys(lst_of_lst[0]))
+    for lst_i in lst_of_lst[1:]:
+        out = [name for name in out if name in lst_i]
+        if len(out) == 0:
+            break
+    return out
+
+
+def union_lists(lst_of_lst: Iterable[Iterable[T]]) -> List[T]:
+    """Performs union of elements in lists (like set union), but keep their original order."""
+    out = {}
+    for lst_i in lst_of_lst:
+        out.update(dict.fromkeys(lst_i))
+    out = list(out)
     return out
 
 
@@ -172,3 +215,42 @@ def flat_list(lst: Iterable[Sequence[T]]) -> Tuple[List[T], List[int]]:
     flatten_lst = [element for sublst in lst for element in sublst]
     sizes = [len(sents) for sents in lst]
     return flatten_lst, sizes
+
+
+@overload
+def unzip(lst: Iterable[Tuple[T]]) -> Tuple[List[T]]:
+    ...
+
+
+@overload
+def unzip(lst: Iterable[Tuple[T, U]]) -> Tuple[List[T], List[U]]:
+    ...
+
+
+@overload
+def unzip(lst: Iterable[Tuple[T, U, V]]) -> Tuple[List[T], List[U], List[V]]:
+    ...
+
+
+@overload
+def unzip(
+    lst: Iterable[Tuple[T, U, V, W]]
+) -> Tuple[List[T], List[U], List[V], List[W]]:
+    ...
+
+
+def unzip(lst):
+    """Invert zip() function.
+
+    .. code-block:: python
+        :caption:  Example
+
+        >>> lst1 = [1, 2, 3, 4]
+        >>> lst2 = [5, 6, 7, 8]
+        >>> lst_zipped = list(zip(lst1, lst2))
+        >>> lst_zipped
+        ... [(1, 5), (2, 6), (3, 7), (4, 8)]
+        >>> unzip(lst_zipped)
+        ... ([1, 2, 3, 4], [5, 6, 7, 8])
+    """
+    return tuple(map(list, zip(*lst)))